\name{bgColor}
\alias{bgColor}
\title{Background color for the chart}
\description{
This specifies the background color of the chart area.
}
\usage{
bgColor
}
\value{
A character string specifying the background color.
}
\examples{
library(yourpackage)
plot(1:5, main = "Example", bg = bgColor)
}
\keyword{datasets}
