% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\docType{data}
\name{bgColor}
\alias{bgColor}
\title{Background color for the chart}
\format{
An object of class \code{character} of length 1.
}
\usage{
bgColor
}
\description{
This specifies the background color of the chart area.
}
\keyword{datasets}
