% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{circle_colors}
\alias{circle_colors}
\title{Colors palette for circles
This palette provides colors suitable for representing circles in charts like pie chart.}
\format{
An object of class \code{character} of length 3.
}
\usage{
circle_colors
}
\description{
Colors palette for circles
This palette provides colors suitable for representing circles in charts like pie chart.
}
\keyword{datasets}
