% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{custom_colors}
\alias{custom_colors}
\title{Custom colors palette for charts}
\format{
An object of class \code{character} of length 5.
}
\usage{
custom_colors
}
\description{
This palette provides custom colors that can be used for chart elements such as series, backgrounds, or legends.
}
\keyword{datasets}
