\name{custom_theme}
\alias{custom_theme}
\title{Tile theme for Highcharter charts}
\description{
This theme provides a customized appearance for Highcharter charts.
}
\usage{
custom_theme()
}
\value{
A highcharter theme object.
}
\examples{
library(yourpackage)
highchart() %>% hc_theme(custom_theme())
}
\keyword{datasets}
