\name{font_family}
\alias{font_family}
\title{Font family for the chart elements}
\description{
This specifies the font family to be used for the chart elements such as titles, axes labels, and legends.
}
\usage{
font_family
}
\value{
A character string specifying the font family.
}
\examples{
library(yourpackage)
plot(1:5, main = "Example", family = font_family)
}
\keyword{datasets}
