\name{line_theme}
\alias{line_theme}
\title{Double line theme for Highcharter charts}
\description{
This theme provides a customized appearance for Highcharter charts with double lines.
}
\usage{
line_theme()
}
\value{
A highcharter theme object.
}
\examples{
library(yourpackage)
highchart() %>% hc_theme(line_theme())
}
\keyword{datasets}
