% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{soft_colors}
\alias{soft_colors}
\title{Colors palette for drilldown charts}
\format{
An object of class \code{character} of length 10.
}
\usage{
soft_colors
}
\description{
This palette provides colors specifically designed for drilldown charts as a soft colors.
}
\keyword{datasets}
