\name{title_font_size}
\alias{title_font_size}
\title{Font size for titles}
\description{
This specifies the font size of the chart titles.
}
\usage{
title_font_size
}
\value{
A character string specifying the font size for titles.
}
\examples{
library(yourpackage)
plot(1:5, main = "Example", cex.main = title_font_size)
}
\keyword{datasets}
