\name{x_axis_font_size}
\alias{x_axis_font_size}
\title{Font size for x-axis labels}
\description{
This specifies the font size of the x-axis labels.
}
\usage{
x_axis_font_size
}
\value{
A character string specifying the font size for x-axis labels.
}
\examples{
library(yourpackage)
plot(1:5, xlab = "X-axis", cex.lab = x_axis_font_size)
}
\keyword{datasets}
