\name{y_axis_font_size}
\alias{y_axis_font_size}
\title{Font size for y-axis labels}
\description{
This specifies the font size of the y-axis labels.
}
\usage{
y_axis_font_size
}
\value{
A character string specifying the font size for y-axis labels.
}
\examples{
library(yourpackage)
plot(1:5, ylab = "Y-axis", cex.lab = y_axis_font_size)
}
\keyword{datasets}
